# task-manager-ms/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:2301",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "TaskManager",
      "description": "The TaskManager API"
    }
  ],
  "paths": {
    "/api/v4/task/{taskId}": {
      "put": {
        "tags": [
          "TaskManager"
        ],
        "summary": "Updates status of the task",
        "operationId": "updateTaskStatus",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "Id of task to be updated",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "New status for task",
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "enum": [
                  "NEW",
                  "IN_PROGRESS",
                  "COMPLETED",
                  "CANCELLED"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestException"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestException"
                }
              }
            }
          },
          "200": {
            "description": "Task`s status is updates"
          }
        }
      },
      "delete": {
        "tags": [
          "TaskManager"
        ],
        "summary": "Deletes the task",
        "operationId": "deleteTask",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "Id of task to be deleted",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestException"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestException"
                }
              }
            }
          },
          "200": {
            "description": "Task is deleted"
          }
        }
      },
      "patch": {
        "tags": [
          "TaskManager"
        ],
        "summary": "Updates a task",
        "operationId": "updateTask",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "Id of task to be updated",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Data for updating a task",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestException"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestException"
                }
              }
            }
          },
          "200": {
            "description": "Task is updated"
          }
        }
      }
    },
    "/api/v4/task": {
      "post": {
        "tags": [
          "TaskManager"
        ],
        "summary": "Creates a new task",
        "operationId": "createTask",
        "requestBody": {
          "description": "Data for creation a new task",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestException"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestException"
                }
              }
            }
          },
          "200": {
            "description": "Task is created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      }
    },
    "/api/v4/task/all": {
      "get": {
        "tags": [
          "TaskManager"
        ],
        "summary": "Returns all tasks",
        "operationId": "getAllTasks",
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestException"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestException"
                }
              }
            }
          },
          "200": {
            "description": "Tasks are returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskDto"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "RestException": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Code of the exception"
          },
          "message": {
            "type": "string",
            "description": "Message of the exception"
          },
          "details": {
            "type": "string",
            "description": "Details of the exception"
          }
        },
        "description": "Response for exception"
      },
      "TaskRequest": {
        "required": [
          "priority",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "description": "status",
            "example": "NEW",
            "enum": [
              "NEW, IN_PROGRESS, COMPLETED, CANCELLED"
            ]
          },
          "priority": {
            "maximum": 10,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "TaskDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "IN_PROGRESS",
              "COMPLETED",
              "CANCELLED"
            ]
          },
          "priority": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}